Updated student's scenarios
The X-distances between the enemies has been updated so now they are
always the same (300) in every scenario but in 04DragonLand.battle where
a Stone is 100 X-units before the Treasure.

1st column of enemies: 300
2nd column of enemies: 600
3rd column of enemies: 900
4th column of enemies: 1100 (only a Stone, in the last scenario)

The students won't be given the race scenarios, which is a disadvantage,
so they will calibrate their Vans based on the distance between enemies,
which will be the same than in the race scenarios.

This will avoid that some students can create perfect Vans for the
practise scenarios but they won't work well in the races.