Removed Booster logic.
It is not working, i.e.:

Two Vans, one with the Booster and one without it and this Fuel
configuration:
		addFuel(CheapPetrol.name, 1000);
		addFuel(ExpensivePetrol.name, 1000);
		addFuel(RacingPetrol.name, 3000);
		(average speed = 1.7)
Van without Booster goes faster.

But sometimes it does work, i.e.:

Two Vans, one with the Booster and one without it and this Fuel
configuration:
		addFuel(CheapPetrol.name, 5000);
		(average speed = 1)
Van with Booster goes faster perfectly.