VanNavigator renamed to VanNavigatorDemo
Why? - It's the easiest way students can implement the VanNavigator
helper class and test it, because...

1) VanHelsing sample uses the VanNavigator class, so the helper class is
included in the folder $RobocodeInstallationFolder/robots/* . Creating a
new projects using this folder&file is tricky and also, students could
modify the behavior of the Animal, Stones, examples (VanNistelroy, etc).

2) Another option could be providing a VanNavigator.java class in
another file, but then, there would be two different VanNavigator.class
binaries generated, it is easy to make a mistake and pick the wrong one.

- VanNavigatorDemo.java will only have 2 methods defined & implemented,
the ones that VanHelsing uses.
- VanNavigator.java will be the class that the students have to
implement and it will have all the methods defined although not all of
them are implemented.