Updated Animal:
Testers reported that sometimes the Van could go through the line of
sight of the Animal and it wouldn't chase the Van. The reason was that
the Animal could go though the defined "isForward" area of the Animal
and before the Animal do a scan, there are 2 reasons for this:

1) Scan arc: Animal do 360 degrees scans, most of the time it's scanning
areas that are not relevant, which gives the Van extra-turns to move
without being caught.

2) isForward area is too small, a bearing of (-2, 2 degrees) is good
enough when the distance is big but when the animal is close it's easy
to avoid. As it is based on degrees, the isForward area (defined by the
arch) will be bigger when the distance is bigger.

Solution:
1) Animal do a initial scan of 45 degrees, and then do a -90 and +90
degrees scan, continuously, so it would be scanning only in a (-45, 45)
range, only in front, it's not scanning the sides or the back. This
increment the scanning frequency in the front area.

2) isForward area is bigger now. When any x-coordinate of the Van
collides with any of the x-coordinate of the Animal, then the Animal
chases the Van. The Van can enter up to 4 pixels in the area of the
Animal without being chased, this is done to visually make more obvious
that the Van is in the line of sight of the Animal. There are
getWidth()-4 = 36-4 = 32 x-positions where the Van is chased.

Updated Animal logic:
- It chases the prey the first time it is scanned.
- It will only kill one Van. The reason is that isForward method only
works if the Animal is heading up or down as it's based on the
x-coordinate, if the Animal is heading any other direction, isForward
won't work in the desired way.
- It prints out information of the scanned robots.